+++ dart/lib/src/http_client/sentry_http_client.dart
@@ -1,7 +1,6 @@
 import 'package:http/http.dart';
+import '../../sentry.dart';
 import 'tracing_client.dart';
-import '../hub.dart';
-import '../hub_adapter.dart';
 import 'breadcrumb_client.dart';
 import 'failed_request_client.dart';