+++ dart/lib/src/hub.dart
@@ -576,6 +576,7 @@ class Hub {
   Future<SentryId> captureTransaction(
     SentryTransaction transaction, {
     SentryTraceContextHeader? traceContext,
+    Hint? hint,
   }) async {
     var sentryId = SentryId.empty();
 
@@ -617,6 +618,7 @@ class Hub {
             transaction,
             scope: item.scope,
             traceContext: traceContext,
+            hint: hint,
           );
         } catch (exception, stackTrace) {
           _options.logger(