+++ dart/lib/src/noop_sentry_client.dart
@@ -67,6 +67,7 @@ class NoOpSentryClient implements SentryClient {
     SentryTransaction transaction, {
     Scope? scope,
     SentryTraceContextHeader? traceContext,
+    Hint? hint,
   }) async =>
       SentryId.empty();