+++ dart/lib/src/protocol/sentry_span.dart
@@ -7,7 +7,10 @@ import '../metrics/local_metrics_aggregator.dart';
 
 import '../sentry_tracer.dart';
 
-typedef OnFinishedCallback = Future<void> Function({DateTime? endTimestamp});
+typedef OnFinishedCallback = Future<void> Function({
+  DateTime? endTimestamp,
+  Hint? hint,
+});
 
 class SentrySpan extends ISentrySpan {
   final SentrySpanContext _context;
@@ -36,6 +39,8 @@ class SentrySpan extends ISentrySpan {
   @override
   final SentryTracesSamplingDecision? samplingDecision;
 
+  late final Hint? hint;
+
   SentrySpan(
     this._tracer,
     this._context,
@@ -55,10 +60,15 @@ class SentrySpan extends ISentrySpan {
   }
 
   @override
-  Future<void> finish({SpanStatus? status, DateTime? endTimestamp}) async {
+  Future<void> finish({
+    SpanStatus? status,
+    DateTime? endTimestamp,
+    Hint? hint,
+  }) async {
     if (finished) {
       return;
     }
+    this.hint = hint;
 
     if (status != null) {
       _status = status;