+++ dart/lib/src/sentry_client.dart
@@ -365,8 +365,9 @@ class SentryClient {
     SentryTransaction transaction, {
     Scope? scope,
     SentryTraceContextHeader? traceContext,
+    Hint? hint,
   }) async {
-    final hint = Hint();
+    hint ??= Hint();
 
     SentryTransaction? preparedTransaction =
         _prepareEvent(transaction, hint) as SentryTransaction;
@@ -409,8 +410,10 @@ class SentryClient {
     preparedTransaction = _createUserOrSetDefaultIpAddress(preparedTransaction)
         as SentryTransaction;
 
-    preparedTransaction =
-        await _runBeforeSend(preparedTransaction, hint) as SentryTransaction?;
+    preparedTransaction = await _runBeforeSend(
+      preparedTransaction,
+      hint,
+    ) as SentryTransaction?;
 
     // dropped by beforeSendTransaction
     if (preparedTransaction == null) {
@@ -515,7 +518,7 @@ class SentryClient {
     try {
       if (event is SentryTransaction && beforeSendTransaction != null) {
         beforeSendName = 'beforeSendTransaction';
-        final callbackResult = beforeSendTransaction(event);
+        final callbackResult = beforeSendTransaction(event, hint);
         if (callbackResult is Future<SentryTransaction?>) {
           processedEvent = await callbackResult;
         } else {