+++ file/test/mock_sentry_client.dart
@@ -17,6 +17,7 @@ class MockSentryClient with NoSuchMethodProvider implements SentryClient {
     SentryTransaction transaction, {
     Scope? scope,
     SentryTraceContextHeader? traceContext,
+    Hint? hint,
   }) async {
     captureTransactionCalls
         .add(CaptureTransactionCall(transaction, traceContext, scope));