+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Features
 
+- Make response body accessible via hint in `beforSend` callback for failed web requests or if tracing is enabled in `SentryHttpClient` ([#2293](https://github.com/getsentry/sentry-dart/pull/2293))
+  ```dart
+  options.beforeSendTransaction = (transaction, hint) {
+    final firstHint = transaction.spans[0].hint;
+    final firstResponseBody = firstHint?.get(TypeCheckHint.httpResponse);
+    final secondHint = transaction.spans[1].hint;
+    final secondResponseBody = secondHint?.get(TypeCheckHint.httpResponse);
+    // user can now use it
+    return transaction;
+  };
+  ```
 - Support for screenshot PII content masking ([#2361](https://github.com/getsentry/sentry-dart/pull/2361))
   By default, masking is enabled for SessionReplay. To also enable it for screenshots captured with events, you can specify `options.experimental.privacy`:
   ```dart